Menambahkan DELETE, & memasukan gambar

+++ Tubes/index.php
@@ -9,15 +9,16 @@ $hewan = query("SELECT * FROM hewan_a");
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Halaman Admin</title>
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
 </head>
 <body>
     
-<h1>Daftar Hewan </h1>
+<h1>Jenis Jenis Hewan </h1>
 
 <a href="tambah.php">Tambah Data Hewan</a>
 <br><br>
 
-<table border="1" cellpadding="10" cellspacing="0">
+<table class = "table-bordered table"  cellpadding="10" cellspacing="0">
     
     <tr>
         <th>No.</th>
@@ -35,8 +36,8 @@ $hewan = query("SELECT * FROM hewan_a");
     <tr>
         <td><?= $i; ?></td>  
         <td>
-            <a href="ubah.php">edit</a> |   
-            <a href="hapus.php">hapus</a>
+            <a class="btn btn-primary" href="edit.php">edit</a> |   
+            <a class="btn btn-primary" href="hapus.php?id=<?= $row ["id"]; ?>" onclick="return confirm('Anda yakin ingin menghapusnya?')">hapus</a>
         </td>
         <td><img src="img/<?= $row ["gambar"]; ?>" width="200   "></td>
         <td><?= $row["nama"]; ?></td>
@@ -47,6 +48,6 @@ $hewan = query("SELECT * FROM hewan_a");
     <?php endforeach; ?>
 
 </table>
-
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p" crossorigin="anonymous"></script>
 </body>
 </html>
